modularize dynamic generics
